Refactored out HeadlessTestBase
