+++ src/test/java/seedu/address/logic/textfieldmodules/HeadlessTestBase.java
@@ -1,0 +1,14 @@
+package seedu.address.logic.textfieldmodules;
+
+import org.junit.jupiter.api.BeforeAll;
+
+public class HeadlessTestBase {
+    @BeforeAll
+    private static void setHeadless() {
+        System.setProperty("testfx.robot", "glass");
+        System.setProperty("testfx.headless", "true");
+        System.setProperty("prism.order", "sw");
+        System.setProperty("prism.text", "t2k");
+        System.setProperty("java.awt.headless", "true");
+    }
+}
